Models Refactoring vol 1
