+++ bin/application.properties
@@ -1,0 +1,17 @@
+spring.datasource.url = jdbc:mysql://localhost:3306/schoolsystemdba2
+
+spring.datasource.username = root
+spring.datasource.password = dariusz
+
+spring.datasource.testWhileIdle = true
+spring.datasource.validationQuery = SELECT 1
+
+spring.jpa.show-sql = true
+
+spring.jpa.hibernate.ddl-auto = update
+
+spring.jpa.hibernate.naming.strategy = org.hibernate.cfg.ImprovedNamingStrategy
+
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+
+spring.jpa.properties.hibernate.current_session_context_class=org.springframework.orm.hibernate4.SpringSessionContext
