--- src/main/java/pl/dn/controller/ByleJakaKlasa.java
@@ -1,7 +1,0 @@
-package pl.dn.controller;
-
-/**
- * Created by User on 21.08.2017.
- */
-public class ByleJakaKlasa {
-}
