--- src/main/java/pl/dn/dao/Y.java
@@ -1,7 +1,0 @@
-package pl.dn.dao;
-
-/**
- * Created by User on 17.08.2017.
- */
-public class Y {
-}
