+++ src/main/java/pl/dn/model/contactInfo/UserContactInfo.java
@@ -1,0 +1,64 @@
+package pl.dn.model.contactInfo;
+
+import javax.persistence.Column;
+import javax.persistence.Embedded;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import pl.dn.model.placeInfo.Address;
+
+@Entity
+@Table(name = "user_contact_info")
+public class UserContactInfo {
+	@Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long id;
+
+    private String email;
+
+    @Column(name = "phone_number")
+    private String phoneNumber;
+
+    @Embedded
+    private Address address;
+
+
+    public UserContactInfo() {
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+
+    public String getPhoneNumber() {
+        return phoneNumber;
+    }
+
+    public void setPhoneNumber(String phoneNumber) {
+        this.phoneNumber = phoneNumber;
+    }
+
+    public Address getAddress() {
+        return address;
+    }
+
+    public void setAddress(Address address) {
+        this.address = address;
+    }
+}
