+++ src/main/java/pl/dn/model/userType/UserLogin.java
@@ -1,0 +1,50 @@
+package pl.dn.model.userType;
+
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+public class UserLogin {
+	@Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private long id;
+
+    private String username;
+    private String password;
+
+    private boolean enabled = true;
+
+	public long getId() {
+		return id;
+	}
+
+	public void setId(long id) {
+		this.id = id;
+	}
+
+	public String getUsername() {
+		return username;
+	}
+
+	public void setUsername(String username) {
+		this.username = username;
+	}
+
+	public String getPassword() {
+		return password;
+	}
+
+	public void setPassword(String password) {
+		this.password = password;
+	}
+
+	public boolean isEnabled() {
+		return enabled;
+	}
+
+	public void setEnabled(boolean enabled) {
+		this.enabled = enabled;
+	}
+    
+    
+}
