--- src/main/java/pl/dn/service/ByleJakaKlasa.java
@@ -1,7 +1,0 @@
-package pl.dn.service;
-
-/**
- * Created by User on 31.08.2017.
- */
-public class ByleJakaKlasa {
-}
